m3u8: workaround for m3u8 live stream
when m3u8 is live stream, is_endlist is False
we need to re-request m3u8 link

but if the download speed is fast, time of two m3u8 link is overlapped

in order to workaround time overlapped, we need to simulate player by using sleep.

but this can't be prefect!

Signed-off-by: Zhang Ning <zhangn1985@gmail.com>